+++ app/[locale]/login/page.tsx
@@ -1,122 +1,226 @@
 'use client'
 
-import React, {useEffect, useRef, useState} from "react";
-import {useRouter} from 'next/navigation';
-import styles from './login.module.scss'
+import React, { useEffect, useRef, useState } from "react";
+import { useRouter } from 'next/navigation';
 import logECS from '@/utils/clientLogger';
-import {useCsrf} from '@/context/CsrfContext';
-import {handleErrors} from '@/utils/errorHandler';
-import {useAuthContext} from '@/context/AuthContext';
+import { useCsrf } from '@/context/CsrfContext';
+import { handleErrors } from '@/utils/errorHandler';
+import { useAuthContext } from '@/context/AuthContext';
+import { useTranslations } from "next-intl";
+import {
+  Button,
+  Form,
+  TextField,
+  Link,
+  Text,
+  Label,
+  Input,
+  FieldError,
+} from "react-aria-components";
+import {
+  LayoutContainer,
+  ContentContainer,
+  ToolbarContainer,
+} from '@/components/Container';
+import styles from './login.module.scss';
 
-type User = {
-    email: string;
-    password: string;
-};
 
-// Placeholder Login page until we get new design
+type LoginSteps =
+  | "email"
+  | "sso"
+  | "password";
+
+
 const LoginPage: React.FC = () => {
-    const [user, setUser] = useState<User>({
-        email: "",
-        password: "",
-    });
-    const [errors, setErrors] = useState<string[]>([]);
-    const [loading, setLoading] = useState(false);
-    const errorRef = useRef<HTMLDivElement>(null);
-    const { csrfToken } = useCsrf();
-    const router = useRouter();
-    const { setIsAuthenticated } = useAuthContext();
+  const t = useTranslations('LoginPage');
 
-    const handleInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-        setUser({ ...user, [event.target.name]: event.target.value });
-    };
+  const formRef = useRef<HTMLFormElement | null>(null);
+  const [step, setStep] = useState<LoginSteps>("email");
 
-    const handleSubmit = async (event: React.FormEvent<HTMLFormElement>) => {
-        event.preventDefault();
-        setLoading(true);
-        setErrors([]); // Clear previous errors
+  const [email, setEmail] = useState("");
+  const [password, setPassword] = useState("");
 
-        const loginRequest = async (token: string | null) => {
-            return await fetch(`${process.env.NEXT_PUBLIC_SERVER_ENDPOINT}/apollo-signin`, {
-                method: 'POST',
-                credentials: 'include',
-                headers: {
-                    'Content-Type': 'application/json',
-                    'X-CSRF-TOKEN': token || '',
-                },
-                body: JSON.stringify(user),
-            });
-        }
+  const [errors, setErrors] = useState<string[]>([]);
+  const [loading, setLoading] = useState(false);
+  const errorRef = useRef<HTMLDivElement>(null);
+  const { csrfToken } = useCsrf();
+  const router = useRouter();
+  const { setIsAuthenticated } = useAuthContext();
 
-        if (csrfToken) {
+  const handleLogin = async () => {
+    setLoading(true);
+    setErrors([]);  // Clear previous errors
 
-            /* eslint-disable @typescript-eslint/no-explicit-any */
-            try {
-                const response = await loginRequest(csrfToken);
+    const loginRequest = async (token: string | null) => {
+      return await fetch(`${process.env.NEXT_PUBLIC_SERVER_ENDPOINT}/apollo-signin`, {
+        method: 'POST',
+        credentials: 'include',
+        headers: {
+          'Content-Type': 'application/json',
+          'X-CSRF-TOKEN': token || '',
+        },
+        body: JSON.stringify({
+          email,
+          password,
+        }),
+      });
+    }
 
-                if (response.ok) {
-                    setIsAuthenticated(true);
-                    router.push('/')
-                } else {
-                    await handleErrors(response, loginRequest, setErrors, router, '/login');
-                }
+    /* eslint-disable @typescript-eslint/no-explicit-any */
+    try {
+      const response = await loginRequest(csrfToken);
 
-            } catch (err: any) {
-                logECS('error', 'Signin error', {
-                    error: err,
-                    url: { path: '/apollo-signin' }
-                });
-            } finally {
-                setLoading(false);
-            }
-        } else {
-            setErrors(prev => prev.concat('Something went wrong'))
-        }
+      if (response.ok) {
+        setIsAuthenticated(true);
+        router.push('/')
+      } else {
+        await handleErrors(response, loginRequest, setErrors, router, '/login');
+      }
 
-    };
+    } catch (err: any) {
+      logECS('error', 'Login error', {
+        error: err,
+        url: { path: '/apollo-signin' }
+      });
+    } finally {
+      setLoading(false);
+    }
+  }
 
-    useEffect(() => {
-        if (errors.length > 0) {
-            if (errorRef.current) {
-                errorRef.current.scrollIntoView({ behavior: 'smooth' });
-                errorRef.current.focus();
-            }
-        }
-    }, [errors])
+  async function handleSubmit(ev: React.FormEvent<HTMLFormElement>) {
+    ev.preventDefault();
 
+    switch (step) {
+      case "email":
+        // TODO: Detect SSO email and redirect to that page. For now just
+        // go to normal login step.
+        setStep("password");
+        break;
 
-    return (
-        <div className={styles.loginWrapper} ref={errorRef}>
-            <form className={styles.loginForm} id="login" onSubmit={handleSubmit}>
-                {errors && errors.length > 0 &&
-                    <div className="error">
-                        {errors.map((error, index) => (
-                            <p key={index}>{error}</p>
-                        ))}
-                    </div>
-                }
-                <h3 className={styles.heading3}>Sign in</h3>
-                <label htmlFor="email">Email</label>
-                <input
-                    type="email"
-                    id="email"
-                    name="email"
-                    value={user.email}
-                    onChange={handleInputChange}
-                    required
-                />
+      case "password":
+        handleLogin();
+        break;
+    }
+  }
 
-                <label htmlFor="password">Password</label>
-                <input
-                    type="password"
-                    id="password"
-                    name="password"
-                    value={user.password}
-                    onChange={handleInputChange}
-                    required
-                />
-                <button type="submit" id="login-button" disabled={loading}>Login</button>
-            </form>
-        </div>
-    );
+  function reFocusForm() {
+    if (formRef.current) {
+      // Focus the first input in the form.
+      const formEl = formRef.current as HTMLElement;
+      const inputEl = formEl.querySelector('input') as HTMLElement;
+      if (inputEl) inputEl.focus();
+    }
+  }
+
+  useEffect(() => {
+    if (errors) {
+      if (errorRef.current) {
+        errorRef.current.scrollIntoView({ behavior: 'smooth' });
+        errorRef.current.focus();
+      }
+    }
+    reFocusForm();
+  }, [errors])
+
+  useEffect(() => {
+    if (step === "password") {
+      if (formRef.current) {
+        const formEl = formRef.current as HTMLElement;
+        const inputEl = formEl.querySelector('#password') as HTMLElement;
+        if (inputEl) inputEl.focus();
+      }
+    }
+  }, [step]);
+
+  return (
+    <LayoutContainer className={styles.loginPage}>
+      <ContentContainer className={styles.loginContent}>
+        <h3>{t('pageTitle')}</h3>
+
+        <Form
+          className={styles.loginForm}
+          onSubmit={handleSubmit}
+          ref={formRef}
+          data-step={step}
+        >
+          {errors && errors.length > 0 &&
+            <div className={`error ${styles.error}`}>
+              {errors.map((error, index) => (
+                <p key={index}>{error}</p>
+              ))}
+            </div>
+          }
+
+          {(step === "email" || "password") && (
+            <TextField
+              name="email"
+              type="email"
+              aria-label={t('emailAddress')}
+              onChange={setEmail}
+              value={email}
+              isRequired
+            >
+              <Label>{t('emailAddress')}</Label>
+              <Input data-testid="emailInput" />
+              {(step === "email") && (
+                <Text slot="description" className={styles.help}>
+                  {t('singleSignOn')}
+                </Text>
+              )}
+              <FieldError />
+            </TextField>
+          )}
+
+          {(step === "password") && (
+            <TextField
+              id="password"
+              name="password"
+              type="password"
+              aria-label={t('password')}
+              onChange={setPassword}
+              isRequired
+            >
+              <Label>{t('password')}</Label>
+              <Input data-testid="passInput" />
+              <FieldError />
+              <Text slot="description" className={styles.help}>
+                <Link>{t('forgotPassword')}</Link>
+              </Text>
+            </TextField>
+          )}
+
+          <ToolbarContainer className={styles.formActions}>
+            {(step === "email") && (
+              <Button
+                type="submit"
+                isDisabled={loading}
+                data-testid="actionContinue"
+              >
+                {loading ? '...' : t('continue')}
+              </Button>
+            )}
+
+            {(step === "password") && (
+              <Button
+                type="submit"
+                isDisabled={loading}
+                data-testid="actionSubmit"
+              >
+                {loading ? '...' : t('login')}
+              </Button>
+            )}
+          </ToolbarContainer>
+
+          <div className={styles.formLinks}>
+            {(step === "email") && (
+              <Link href="/signup/">{t('signupText')}</Link>
+            )}
+            <Link href="#">{t('getHelpText')}</Link>
+          </div>
+        </Form>
+      </ContentContainer>
+    </LayoutContainer>
+  );
 };
+
 export default LoginPage;